Handle special characters in passwords
This change only addresses passwords in Pollers and Defaults. The bigger refactor is to use HarvestConfig through out the codebase, but that was too big a change at the moment. That change touches a lot more code.

When that change is made, the code in conf.LoadConfig can be removed.