Infinite sleep without needing a TTY or side effects
This commit changes the means by which the container keeps itself
alive from running a `sh` process, to an explicit infinite sleep.

Using a shell means the container does not continue to run after
poller startup if run in detached mode, since stdin is closed the
shell process immediately exits. When run in attached mode,
use of shell will interpret any stdin as commands to be run within
the container, which is potentially dangerous, and certainly unexpected.

Fixes #164