fix: harvest_conf should be consulted when loading files from conf
directory

Fixes #368

This change addresses the problem when using `bin/poller` and `bin/grafana`,
but does not work with `bin/harvest`.

The issue with `bin/harvest` is the `os.Exec`-ed executables are not found
without more directory work.

The workaround is `cd` into the installation directory before executing
`bin/harvest`