fix: first autosupport runs after delay
If the schedule is once a week, the first time autosupport runs, it instead runs after two weeks.

time |---X----f----y----|
         ^    ^
         |    |
   now --     ---- future

That's because when a task is created and added to the schedule at X, a time in the future is calculated (current time + 1 week), at f.

Later, when a go routine checks if the future task is due to run, it determines that by calculating: one week - time.Since(f), but since f is already a week in the future. The task won't run until y, which is two weeks after X was scheduled.

The fix: when the task is scheduled, schedule it at the current time instead of current time + 1 week.