perf: reduce the memory footprint of REST collector (#3303)
* perf: reduce the memory footprint of REST collector

When the REST collector walks pagination links, it does so recursively. Each of these recursive calls retains the previous REST response on the stack, and when there are many recursive calls, a lot of memory can be used. Replace the recursive calls with iterative ones.

* perf: reduce the memory footprint of REST collector

When the REST collector walks pagination links, it does so recursively. Each of these recursive calls retains the previous REST response on the stack, and when there are many recursive calls, a lot of memory can be used. Replace the recursive calls with iterative ones.