perf: improve memory and cpu performance of rest collector
Using the gjson `String()` method will often cause the entire JSON string to be retained in memory.

This PR copies gjson to the thirdparty package and adds a `ClonedString` method to `gjson.Result`. The `ClonedString` method returns a cloned copy of the Result string so that the larger JSON string is available for garbage collection.

The Rest and RestPerf collectors were updated to use `gjson.ClonedString()`.

A similar fix was made in #2054 for RestPerf.